refactored for easier usage + added extra items handling strategy in domain
