+++ disma-cli/src/infra/config/mod.rs
@@ -1,6 +1,5 @@
 pub mod category;
 pub mod channel;
 pub mod guild;
-pub mod options;
 pub mod permission;
 pub mod role;
