+++ disma/src/domain/diff/mod.rs
@@ -2,5 +2,4 @@ pub mod base;
 pub mod category;
 pub mod channel;
 pub mod event;
-pub mod factory;
 pub mod roles;
