+++ disma/src/domain/entities/category/mod.rs
@@ -1,0 +1,11 @@
+mod awaiting;
+pub use awaiting::*;
+
+pub use base::*;
+mod base;
+
+pub use diff::*;
+mod diff;
+
+pub use existing::*;
+mod existing;
