+++ disma/src/domain/entities/channel/mod.rs
@@ -1,0 +1,11 @@
+mod awaiting;
+pub use awaiting::*;
+
+mod base;
+pub use base::*;
+
+mod diff;
+pub use diff::*;
+
+mod existing;
+pub use existing::*;
