+++ disma/src/domain/entities/role/mod.rs
@@ -1,0 +1,11 @@
+mod awaiting;
+pub use awaiting::*;
+
+mod base;
+pub use base::*;
+
+mod existing;
+pub use existing::*;
+
+mod diff;
+pub use diff::*;
